# modules/devconsole/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gorbital dev console",
    "version": "1",
    "description": "Development-only APIs of a gorbital app (ADR-0065), served when APP_ENV is development and DEV_CONSOLE_TOKEN is set; orb dev sets a new token on every run and prints it. Every request needs a Host header of localhost, 127.0.0.1 or [::1] with the app's port, a connection from a loopback address, and Authorization: Bearer <token>. Responses carry no CORS headers and say Cache-Control: no-store. Fields may be added."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "security": [
    {
      "devConsoleToken": []
    }
  ],
...
  "paths": {
...
    "/_dev/requests/stream": {
      "get": {
        "operationId": "dev-requests-stream",
        "summary": "Live requests",
        "description": "Server-Sent Events: `retry: 3000`, then `event: request` with a Request for each new item, `event: dropped` with a StreamDropped when the client was too slow, `: keep-alive` comments every 15 seconds, and a final `event: end` with a StreamEnd after 30 minutes or when the app shuts down. EventSource can't send the Authorization header: read the stream with fetch. At most 8 streams at once.",
        "tags": [
          "Dev console"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong token (code unauthorized)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host header isn't localhost, 127.0.0.1 or [::1] with the app's port, or the connection isn't local (code forbidden)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "This app doesn't serve the endpoint (code not_found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "The console already serves its maximum number of streams (code rate_limited)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/_dev/logs/stream": {
      "get": {
        "operationId": "dev-logs-stream",
        "summary": "Live log records",
        "description": "Server-Sent Events: `retry: 3000`, then `event: log` with a Log for each new item, `event: dropped` with a StreamDropped when the client was too slow, `: keep-alive` comments every 15 seconds, and a final `event: end` with a StreamEnd after 30 minutes or when the app shuts down. EventSource can't send the Authorization header: read the stream with fetch. At most 8 streams at once.",
        "tags": [
          "Dev console"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong token (code unauthorized)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Host header isn't localhost, 127.0.0.1 or [::1] with the app's port, or the connection isn't local (code forbidden)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "This app doesn't serve the endpoint (code not_found)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "The console already serves its maximum number of streams (code rate_limited)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "devConsoleToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token orb dev prints (DEV_CONSOLE_TOKEN)"
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "code"
        ]
      },
...
    }
  }
}